Some cleanup - mostly in the vobsub writer

+++ src/Logic/TransportStream/Helper.cs
@@ -290,7 +290,7 @@ namespace Nikse.SubtitleEdit.Logic.TransportStream
 
         public static int GetLittleEndian32(byte[] buffer, int index)
         {
-            return ((int)buffer[index + 3] << 24 | (int)buffer[index + 2] << 16 | (int)buffer[index + 1] << 8 | (int)buffer[index + 0]);
+            return (buffer[index + 3] << 24 | (int)buffer[index + 2] << 16 | (int)buffer[index + 1] << 8 | (int)buffer[index + 0]);
         }
